rent_golem/backend/rent_golem/accounts/serializers.py
from rest_framework import serializers
from .models import User


class UserSerializer(serializers.ModelSerializer):
    class Meta:
        model = User
        fields = ['email', 'service_password', 'is_staff', 'is_active', 'date_joined']

