rent_golem/backend/requirements.txt
Django==4.2.4
djangorestframework==3.14.0
psycopg==3.1.9
celery==5.3.1
django-celery-results==2.5.1
django-celery-beat==2.5.0
redis==4.6.0
yapapi==0.11.0
django-extensions==3.2.3
djangorestframework-simplejwt==5.3.0
djangorestframework-simplejwt[crypto]